+++ src/crypto/test/crypto.cpp
@@ -3,6 +3,7 @@
 
 #define DOCTEST_CONFIG_IMPLEMENT_WITH_MAIN
 #include "ccf/crypto/base64.h"
+#include "ccf/crypto/eddsa_key_pair.h"
 #include "ccf/crypto/entropy.h"
 #include "ccf/crypto/hmac.h"
 #include "ccf/crypto/jwk.h"
@@ -92,6 +93,17 @@ TEST_CASE("Sign, verify, with KeyPair")
   }
 }
 
+TEST_CASE("Sign, verify, with KeyPair of EdDSA")
+{
+  constexpr auto curve = "curve25519";
+  constexpr auto curve_id = CurveID::CURVE25519;
+  INFO("With curve: " << curve);
+  auto kp = make_eddsa_key_pair(curve_id);
+  vector<uint8_t> contents(contents_.begin(), contents_.end());
+  const vector<uint8_t> signature = kp->sign(contents);
+  CHECK(kp->verify(contents, signature));
+}
+
 TEST_CASE("Sign, verify, with PublicKey")
 {
   for (const auto curve : supported_curves)
@@ -105,6 +117,20 @@ TEST_CASE("Sign, verify, with PublicKey")
     auto pubk = make_public_key(public_key);
     CHECK(pubk->verify(contents, signature));
   }
+}
+
+TEST_CASE("Sign, verify, with PublicKey of EdDSA")
+{
+  constexpr auto curve = "curve25519";
+  constexpr auto curve_id = CurveID::CURVE25519;
+  INFO("With curve: " << curve);
+  auto kp = make_eddsa_key_pair(curve_id);
+  vector<uint8_t> contents(contents_.begin(), contents_.end());
+  const vector<uint8_t> signature = kp->sign(contents);
+
+  const auto public_key = kp->public_key_pem();
+  auto pubk = make_eddsa_public_key(public_key);
+  CHECK(pubk->verify(contents, signature));
 }
 
 TEST_CASE("Sign, fail to verify with bad signature")
@@ -123,6 +149,21 @@ TEST_CASE("Sign, fail to verify with bad signature")
   }
 }
 
+TEST_CASE("Sign, fail to verify with bad signature (EdDSA)")
+{
+  constexpr auto curve = "curve25519";
+  constexpr auto curve_id = CurveID::CURVE25519;
+  INFO("With curve: " << curve);
+  auto kp = make_eddsa_key_pair(curve_id);
+  vector<uint8_t> contents(contents_.begin(), contents_.end());
+  vector<uint8_t> signature = kp->sign(contents);
+
+  const auto public_key = kp->public_key_pem();
+  auto pubk = make_eddsa_public_key(public_key);
+  corrupt(signature);
+  CHECK_FALSE(pubk->verify(contents, signature));
+}
+
 TEST_CASE("Sign, fail to verify with bad contents")
 {
   for (const auto curve : supported_curves)
@@ -137,6 +178,21 @@ TEST_CASE("Sign, fail to verify with bad contents")
     corrupt(contents);
     CHECK_FALSE(pubk->verify(contents, signature));
   }
+}
+
+TEST_CASE("Sign, fail to verify with bad contents (EdDSA)")
+{
+  constexpr auto curve = "curve25519";
+  constexpr auto curve_id = CurveID::CURVE25519;
+  INFO("With curve: " << curve);
+  auto kp = make_eddsa_key_pair(curve_id);
+  vector<uint8_t> contents(contents_.begin(), contents_.end());
+  vector<uint8_t> signature = kp->sign(contents);
+
+  const auto public_key = kp->public_key_pem();
+  auto pubk = make_eddsa_public_key(public_key);
+  corrupt(contents);
+  CHECK_FALSE(pubk->verify(contents, signature));
 }
 
 TEST_CASE("Sign, fail to verify with wrong key on correct curve")
@@ -154,6 +210,23 @@ TEST_CASE("Sign, fail to verify with wrong key on correct curve")
     CHECK_FALSE(pubk->verify(contents, signature));
   }
 }
+
+TEST_CASE("Sign, fail to verify with wrong key on correct curve (EdDSA)")
+{
+    constexpr auto curve = "curve25519";
+    constexpr auto curve_id = CurveID::CURVE25519;
+    INFO("With curve: " << curve);
+    auto kp = make_eddsa_key_pair(curve_id);
+    vector<uint8_t> contents(contents_.begin(), contents_.end());
+    vector<uint8_t> signature = kp->sign(contents);
+
+    auto kp2 = make_eddsa_key_pair(curve_id);
+    const auto public_key = kp2->public_key_pem();
+    auto pubk = make_eddsa_public_key(public_key);
+    CHECK_FALSE(pubk->verify(contents, signature));
+}
+
+// MYTODO: test non supported curve
 
 TEST_CASE("Sign, fail to verify with wrong key on wrong curve")
 {